+++ lib/ClangImporter/ImporterImpl.h
@@ -17,6 +17,7 @@
 #ifndef SWIFT_CLANG_IMPORTER_IMPL_H
 #define SWIFT_CLANG_IMPORTER_IMPL_H
 
+#include "ImportEnumInfo.h"
 #include "SwiftLookupTable.h"
 #include "swift/ClangImporter/ClangImporter.h"
 #include "swift/AST/ASTContext.h"
@@ -246,25 +247,6 @@ class LLVM_LIBRARY_VISIBILITY ClangImporter::Implementation
   friend class SwiftNameLookupExtension;
 
 public:
-  /// \brief Describes how a particular C enumeration type will be imported
-  /// into Swift. All of the possibilities have the same storage
-  /// representation, but can be used in different ways.
-  enum class EnumKind {
-    /// \brief The enumeration type should map to an enum, which means that
-    /// all of the cases are independent.
-    Enum,
-    /// \brief The enumeration type should map to an option set, which means that
-    /// the constants represent combinations of independent flags.
-    Options,
-    /// \brief The enumeration type should map to a distinct type, but we don't
-    /// know the intended semantics of the enum constants, so conservatively
-    /// map them to independent constants.
-    Unknown,
-    /// \brief The enumeration constants should simply map to the appropriate
-    /// integer values.
-    Constants
-  };
-
   Implementation(ASTContext &ctx, const ClangImporterOptions &opts);
   ~Implementation();
 
@@ -519,13 +501,38 @@ private:
   /// \brief Cache of the class extensions.
   llvm::DenseMap<ClassDecl *, CachedExtensions> ClassExtensions;
 
+  /// \brief Cache enum infos
+  llvm::DenseMap<const clang::EnumDecl *, importer::EnumInfo> enumInfos;
+
 public:
   /// \brief Keep track of subscript declarations based on getter/setter
   /// pairs.
   llvm::DenseMap<std::pair<FuncDecl *, FuncDecl *>, SubscriptDecl *> Subscripts;
 
-  /// \brief Keep track of enum constant name prefixes in enums.
-  llvm::DenseMap<const clang::EnumDecl *, StringRef> EnumConstantNamePrefixes;
+  importer::EnumInfo getEnumInfo(const clang::EnumDecl *decl,
+                                 clang::Preprocessor *ppOverride = nullptr) {
+    if (enumInfos.count(decl))
+      return enumInfos[decl];
+    // Due to the semaOverride present in importFullName(), we might be using a
+    // decl from a different context.
+    auto &preprocessor = ppOverride ? *ppOverride : getClangPreprocessor();
+
+    importer::EnumInfo enumInfo(decl, preprocessor);
+    enumInfos[decl] = enumInfo;
+    return enumInfo;
+  }
+  importer::EnumKind getEnumKind(const clang::EnumDecl *decl,
+                                 clang::Preprocessor *ppOverride = nullptr) {
+    return getEnumInfo(decl, ppOverride).getKind();
+  }
+
+  /// \brief the prefix to be stripped from the names of the enum constants
+  /// within the given enum.
+  StringRef
+  getEnumConstantNamePrefix(const clang::EnumDecl *decl,
+                            clang::Preprocessor *ppOverride = nullptr) {
+    return getEnumInfo(decl, ppOverride).getConstantNamePrefix();
+  }
 
 private:
   class EnumConstantDenseMapInfo {
@@ -546,11 +553,6 @@ private:
       return lhs == rhs;
     }
   };
-
-  /// Retrieve the prefix to be stripped from the names of the enum constants
-  /// within the given enum.
-  StringRef getEnumConstantNamePrefix(clang::Sema &sema,
-                                      const clang::EnumDecl *enumDecl);
 
 public:
   /// \brief Keep track of enum constant values that have been imported.
@@ -908,11 +910,6 @@ public:
 
   /// Returns true if it is expected that the macro is ignored.
   bool shouldIgnoreMacro(StringRef name, const clang::MacroInfo *macro);
-
-  /// \brief Classify the given Clang enumeration type to describe how it
-  /// should be imported 
-  static EnumKind classifyEnum(clang::Preprocessor &pp,
-                               const clang::EnumDecl *decl);
 
   /// Import attributes from the given Clang declaration to its Swift
   /// equivalent.
